HSEARCH-3759 Move the schema validation context to the engine
We're going to need it for Lucene, too. And probably in the mapper
later.
