+++ engine/src/main/java/org/hibernate/search/engine/reporting/impl/EngineEventContextMessages.java
@@ -60,6 +60,9 @@ public interface EngineEventContextMessages {
 	@Message(value = "shard '%1$s'")
 	String shard(String shardId);
 
+	@Message(value = "schema validation")
+	String schemaValidation();
+
 	@Message(value = "index schema root")
 	String indexSchemaRoot();
